Isolate bundle RPC implementation
